Organize grundsteuer input model into section-specific files.
