+++ erica/request_processing/erica_input/v2/possibly_aliased_input.py
@@ -1,0 +1,8 @@
+from humps import camelize
+from pydantic import BaseModel
+
+
+class PossiblyAliasedInput(BaseModel):
+    class Config:
+        alias_generator = camelize
+        allow_population_by_field_name = True
